requirements.txt
streamlit
pandas
requests
odfpy
python-dotenv
